f2fs: handle data and node segment in same section
In cases where active_logs is specified as less than 6 during mount time,
it was observed that both data segments and node segments can be allocated
from the same section. But do_garbage_collect() assumes one type of segments
for an entire section and this can cause false inconsistency error conditions.
do_garbage_collect() assumes the first seg type to be same to all segments
in the section and compares the segment_type with the summary type of
remaining segments causing false error.
Get each segments type and compare it with its summary type to ensure that
there are no inconsistencies.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>